update table orders display

+++ routes/user.js
@@ -1,5 +1,5 @@
 //---------------------------------------------signup page call------------------------------------------------------
-exports.signup = function (req, res) {
+exports.signup = function(req, res) {
   message = "";
   if (req.method == "POST") {
     var post = req.body;
@@ -25,7 +25,7 @@ exports.signup = function (req, res) {
       pass +
       "')";
 
-    var query = db.query(sql, function (err, result) {
+    var query = db.query(sql, function(err, result) {
       message = "Succesfully! Your account has been created.";
       res.render("signup.ejs", { message: message });
     });
@@ -35,7 +35,7 @@ exports.signup = function (req, res) {
 };
 
 //-----------------------------------------------login page call------------------------------------------------------
-exports.login = function (req, res) {
+exports.login = function(req, res) {
   var message = "";
   var sess = req.session;
 
@@ -53,7 +53,7 @@ exports.login = function (req, res) {
       pass +
       "'";
 
-    db.query(sql, function (err, results) {
+    db.query(sql, function(err, results) {
       if (results.length) {
         req.session.userId = results[0].id;
         req.session.first_name = results[0].first_name;
@@ -67,14 +67,13 @@ exports.login = function (req, res) {
         res.render("index.ejs", { message: message });
       }
     });
-
   } else {
     res.render("index.ejs", { message: message });
   }
 };
 //-----------------------------------------------dashboard page functionality----------------------------------------------
 
-exports.dashboard = function (req, res, next) {
+exports.dashboard = function(req, res, next) {
   var user = req.session.user,
     userId = req.session.userId,
     fname = req.session.first_name;
@@ -89,26 +88,26 @@ exports.dashboard = function (req, res, next) {
 
   var sql = "SELECT * FROM `elunch_users2` WHERE `id`='" + userId + "'";
 
-  db.query(sql, function (err, results) {
+  db.query(sql, function(err, results) {
     res.render("dashboard.ejs", { fname });
     // console.log("user= " + fname);
   });
 };
 //------------------------------------logout functionality----------------------------------------------
-exports.logout = function (req, res) {
-  req.session.destroy(function (err) {
+exports.logout = function(req, res) {
+  req.session.destroy(function(err) {
     res.redirect("/login");
   });
 };
 //--------------------------------render user details after login--------------------------------
-exports.profile = function (req, res) {
+exports.profile = function(req, res) {
   var userId = req.session.userId;
   if (userId == null) {
     res.redirect("/login");
     return;
   }
   var sql = "SELECT * FROM `elunch_users2` WHERE `id`='" + userId + "'";
-  db.query(sql, function (err, result) {
+  db.query(sql, function(err, result) {
     res.render("profile.ejs", { result });
   });
 };
@@ -127,7 +126,7 @@ exports.profile = function (req, res) {
 // };
 //-----------------------------------------------new_user page functionality----------------------------------------------
 
-exports.new_order = function (req, res, next) {
+exports.new_order = function(req, res, next) {
   var user = req.session.user,
     userId = req.session.userId,
     fname = req.session.first_name,
@@ -135,65 +134,72 @@ exports.new_order = function (req, res, next) {
     menu_json,
     menu_json2,
     mydate,
-    my_orders, order_no;
+    my_orders,
+    order_no, mydate3;
   var sesa_no2 = req.body.sesa_no1;
   var supplier_post = req.body.supplier;
   var order_date = req.body.order_date;
   var order_no5 = req.body.order_no;
   var menu_desctription, menu_price;
   var supplier;
-  mydate=req.query.mydate;
+  mydate = req.query.mydate;
   console.log("Method: " + req.method);
-  console.log("mydate: "+ mydate);
-  sesa_no1="48177";
-  // if (userId == null) {
-  //   res.redirect("/login");
-  //   return;
-  // }
+  console.log("mydate: " + mydate);
+  sesa_no1 = "48177";
+
+  var parts2 = mydate.split('-');
+  console.log (parts2[0], parts2[1],  parts2[2]);
+  mydate3 = mydate3 + "-"+ parts2[1] + "-" + parts2[2] ;
+  console.log ("mydate3" + mydate3);
+
+  if (userId == null) {
+    res.redirect("/login");
+    return;
+  }
 
   if (req.method == "DELETE") {
     var delete_id = req.body.delete_id;
     console.log("delete: ", delete_id);
     var sql7 = "DELETE FROM `elunch_orders2` WHERE `id` ='" + delete_id + "'";
-    db.query(sql7, function (err, results) {
+    db.query(sql7, function(err, results) {
       console.log("deleted one row");
     });
   } //delete
 
   if (req.method == "POST") {
-    
     var supplier = req.body.supplier;
-    console.log("supplier w post: ",supplier);
+    console.log("supplier w post: ", supplier);
     var order_date = req.body.order_date;
     var order_no = req.body.order_no;
     console.log(
-      "post: ", sesa_no1, " ",
-      supplier, " ",
-      order_date, " ",
-      "Numer dania ", order_no, " "
+      "post: ",
+      sesa_no1,
+      " ",
+      supplier,
+      " ",
+      order_date,
+      " ",
+      "Numer dania ",
+      order_no,
+      " "
     );
-    mydate=order_date;
-
+    mydate = order_date;
   } else if (req.method == "GET") {
     var supplier = req.query.supplier;
     var order_no = req.body.order_no;
     if (!supplier) {
       supplier = "Mucha"; // Initial value
-    };
-    if (mydate == "undefined"){
-      mydate=req.query.mydate;
-      if (!mydate){
-        mydate="2020-01-26";
-      } 
     }
-    console.log(
-      "get: ", sesa_no1, " ",
-      supplier, " ", mydate
-
-    );
+    if (mydate == "undefined") {
+      mydate = req.query.mydate;
+      if (!mydate) {
+        mydate = "2020-01-26";
+      }
+    }
+    console.log("get: ", sesa_no1, " ", supplier, " ", mydate);
   }
 
-  //supplier="Mucha";
+  
   if (order_no) {
     //base order_no, get from menu order_name and order_price
     var sql6 =
@@ -203,7 +209,7 @@ exports.new_order = function (req, res, next) {
       order_no +
       "'";
 
-    db.query(sql6, function (err, results) {
+    db.query(sql6, function(err, results) {
       if (results.length) {
         menu_desctription = results[0].menu_desctription;
         console.log("menu_desctription: ", menu_desctription);
@@ -211,39 +217,50 @@ exports.new_order = function (req, res, next) {
         //return menu_desctription;
         console.log("menu_price: ", menu_price);
       } else {
-        message = "problem z pobraniem danych z bazy menu przed zapisem do bazy order2";
+        message =
+          "problem z pobraniem danych z bazy menu przed zapisem do bazy order2";
         res.render("index.ejs", { message: message });
       }
       //console.log("rezultat", rezultat);
       // console.log("\n po if order_price2 z sesji: ", req.session.menu_price2 ," \n order_name2 z sesji : ",req.session.order_name2);
       console.log("po if menu_price: ", menu_price);
       console.log("po if menu_desctription", menu_desctription);
-      console.log("supplier: ",supplier);
+      console.log("supplier: ", supplier);
       // put order to DB
       var sql5 =
         "INSERT INTO `elunch_orders2`(`Id_sesa_no`,`order_date`,`order_supplier_name`,`order_no`,`order_name`, `order_price`) VALUES ('" +
-        sesa_no2 + "','" + order_date + "','" + supplier + "','" + order_no5 + "','" + menu_desctription + "','" + menu_price + "')";
+        sesa_no2 +
+        "','" +
+        order_date +
+        "','" +
+        supplier +
+        "','" +
+        order_no5 +
+        "','" +
+        menu_desctription +
+        "','" +
+        menu_price +
+        "')";
 
-      db.query(sql5, function (err, results) {
+      db.query(sql5, function(err, results) {
         console.log("menu_price po insert ", menu_price);
         console.log("menu_desctription po insert  ", menu_desctription);
         console.log("Inerted record to DB");
       });
-
     });
-
   }
+  
+  // Display current orders by date
+  var sql1 =
+    //  "SELECT * FROM `elunch_orders2` WHERE 1" ;
+    //  var sql1 =
+      "SELECT * FROM `elunch_orders2` WHERE `order_date`='" + mydate3 +"' ORDER BY id DESC" ;
+    // var sql1 =
+    //  "SELECT * FROM `elunch_orders2` WHERE `order_date`='" + mydate + "'";
 
-  // Display current orders by date 
-  // var sql1 =
-  //   "SELECT * FROM `elunch_orders2` WHERE 1" ;
-  //  var sql1 =
-  //  "SELECT * FROM `elunch_orders2` WHERE `order_date`='" + mydate +"' ORDER BY id DESC" ;
-   var sql1 =
-   "SELECT * FROM `elunch_orders2` WHERE `order_date`='" + mydate +"'";
    console.log(sql1);
 
-  db.query(sql1, function (err, results) {
+  db.query(sql1, function(err, results) {
     my_orders = JSON.stringify(results);
     console.log("my_orders: ", my_orders);
   });
@@ -257,7 +274,7 @@ exports.new_order = function (req, res, next) {
     // "' ORDER BY id DESC '" +
     "'";
 
-  db.query(sql3, function (err, results) {
+  db.query(sql3, function(err, results) {
     menu_json = JSON.stringify(results);
     // menu_json2 = JSON.stringify(results);
     console.log("menu_json: ", menu_json);
@@ -270,7 +287,7 @@ exports.new_order = function (req, res, next) {
     mysupplier_name +
     "'";
 
-  db.query(sql4, function (err, results) {
+  db.query(sql4, function(err, results) {
     menu_json2 = JSON.stringify(results);
     console.log("menu_json2: ", menu_json2);
     res.render("new_order.ejs", {
@@ -281,13 +298,14 @@ exports.new_order = function (req, res, next) {
       my_orders
     });
   });
-
 };
 //--------------------------------render user details after login--------------------------------
-exports.orders = function (req, res) {
+exports.orders = function(req, res) {
   var userId = req.session.userId,
     fname = req.session.first_name,
-    sesa_no1 = req.session.sesa_no1, data_from, data_to;
+    sesa_no1 = req.session.sesa_no1,
+    data_from,
+    data_to;
 
   data_from = req.body.data_from;
   data_to = req.body.data_to;
@@ -300,15 +318,12 @@ exports.orders = function (req, res) {
   // }
 
   if (req.method == "POST") {
-
-
     var sql = "SELECT * FROM `elunch_orders2` WHERE `id`='" + userId + "'";
     // db.query(sql, function(err, result) {
     //   menu_orders = JSON.stringify(results);
     //   console.log("menu_orders: ", menu_orders);
     //   res.render("orders.ejs", {menu_orders});
     // });
-
   }
   res.render("orders.ejs");
 };
